pyproject.toml
[tool.poetry]
name = "market"
version = "0.1.0"
description = "Yandex Market Order Bot"
authors = ["saikonohack <saintklovus@gmail.com>"]
readme = "README.md"
packages = [{include = "bot"}]

[tool.poetry.dependencies]
python = ">=3.10"
aiogram = "^3.13.1"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
colorlog = "^6.8.2"
redis = "^5.2.1"
tenacity = "^9.0.0"
pytest = "^8.3.5"
prometheus-client = "^0.21.1"
babel = "^2.17.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"